backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.48.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.4
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
